+++ R/growthModels.R
@@ -35,8 +35,9 @@
 #'   }
 #'   \item Richards
 #'   \itemize{
-#'     \item Within FSA, Linf is the mean asymptotic length, ti is the age at the inflection point, k controls the slope at the inflection point (maximum relative growth rate), b is dimensionless but related to the vertical position (i.e., size) of the inflection point, a is dimensionless but related to the horizontal position (i.e., age) of the inflection point, and L0 is the mean length at age-0.
-#'     \item The parameterizations (1-6) correspond to functions/equations 1, 4, 5, 6, 7, and 8, respectively, in Tjorve and Tjorve (2010). Note that their A, S, k, d, and B are Linf, a, k, b, and L0, respectively, here (in FSA). Their (Tjorve and Tjorve 2010) K does not appear here.
+#'     \item Only 4-parameter parameterizations from Tjorve and Tjorve (2010) that seemed useful for modeling fish growth are provided here.
+#'     \item Within FSA, Linf is the mean asymptotic length; ti is the age at the inflection point; k controls the slope at the inflection point (maximum relative growth rate); a is dimensionless but related to the horizontal position (i.e., age) of the inflection point; b, c, and d are dimensionless but related to the vertical position (i.e., size) of the inflection point; and L0 is the mean length at age-0.
+#'     \item The parameterizations (1-5) correspond to functions/equations 5, 3(alt), 7, 4, and 6, respectively, in Tjorve and Tjorve (2010). Note that their A, S, k are Linf, a, k and their d is b, c, and d, respectively, here (in FSA).
 #'   }
 #'   \item logistic
 #'   \itemize{
@@ -123,19 +124,17 @@
 #' ( gompT <- GompertzFuns("Troynikov1"))
 #'
 #' ## Simple Examples -- Richards
-#' ( rich1 <- RichardsFuns() )
-#' plot(rich1(ages,Linf=800,k=0.5,a=1,b=6)~ages,type="b",pch=19)
+#' ( rich1 <- RichardsFuns(1) )
+#' plot(rich1(ages,Linf=800,k=0.5,ti=3,b1=0.15)~ages,type="b",pch=19)
 #' ( rich2 <- RichardsFuns(2) )
-#' plot(rich2(ages,Linf=800,k=0.5,ti=3,b=6)~ages,type="b",pch=19)
+#' plot(rich2(ages,Linf=800,k=0.5,t0=-1,b2=6)~ages,type="b",pch=19)
 #' ( rich3 <- RichardsFuns(3) )
-#' plot(rich3(ages,Linf=800,k=0.5,ti=3,b=0.15)~ages,type="b",pch=19)
+#' plot(rich3(ages,Linf=800,k=0.5,L0=50,b3=1.5)~ages,type="b",pch=19)
 #' ( rich4 <- RichardsFuns(4) )
-#' plot(rich4(ages,Linf=800,k=0.5,ti=3,b=0.95)~ages,type="b",pch=19)
-#' lines(rich4(ages,Linf=800,k=0.5,ti=3,b=1.5)~ages,type="b",pch=19,col="blue")
+#' plot(rich4(ages,Linf=800,k=0.5,ti=3,b2=6)~ages,type="b",pch=19)
 #' ( rich5 <- RichardsFuns(5) )
-#' plot(rich5(ages,Linf=800,k=0.5,L0=50,b=1.5)~ages,type="b",pch=19)
-#' ( rich6 <- RichardsFuns(6) )
-#' plot(rich6(ages,Linf=800,k=0.5,ti=3,Lninf=50,b=1.5)~ages,type="b",pch=19)
+#' plot(rich5(ages,Linf=800,k=0.5,ti=3,b3=0.95)~ages,type="b",pch=19)
+#' lines(rich5(ages,Linf=800,k=0.5,ti=3,b3=1.5)~ages,type="b",pch=19,col="blue")
 #' ( rich2c <- RichardsFuns(2,simple=TRUE) ) # compare to rich2
 #' 
 #' ## Simple Examples -- Logistic
@@ -179,6 +178,7 @@
 #' summary(fit3,correlation=TRUE)
 #' curve(vb3(x,L1=coef(fit3),t1=c(0,4)),from=0,to=5,col="green",lwd=2,add=TRUE)
 #'
+#'
 #' ## Gompertz
 #' # Make some fake data using the original parameterization
 #' gompO <- GompertzFuns("original")
@@ -213,62 +213,9 @@
 #' curve(gomp3(x,Linf=coef(fit3)[1],gi=coef(fit3)[2],t0=coef(fit3)[3]),
 #'       from=0,to=15,col="green",lwd=2,add=TRUE)
 #' 
-#' ## Richards
-#' 
-#' \dontrun{
-#' # Fit first Richards parameterization ... DOES NOT CONVERGE
-#' fit1 <- nls(len~rich1(age,Linf,k,a,b),data=df,
-#'             start=list(Linf=450,k=0.3,a=0.2,b=3))
-#' summary(fit1,correlation=TRUE)
-#' curve(rich1(x,Linf=coef(fit1)),from=0,to=15,col="red",lwd=10,add=TRUE)
-#'
-#' # Fit second Richards parameterization ... DOES NOT CONVERGE
-#' fit2 <- nls(len~rich2(age,Linf,k,ti,b),data=df,
-#'             start=list(Linf=450,k=0.25,ti=3,b=3))
-#' summary(fit2,correlation=TRUE)
-#' curve(rich2(x,Linf=coef(fit2)),from=0,to=15,col="blue",lwd=7,add=TRUE)
-#' }
-#'
-#' plot(len~age,data=df,pch=19,col=rgb(0,0,0,1/5))
-#' 
-#' # Fit third Richards parameterization
-#' fit3 <- nls(len~rich3(age,Linf,k,ti,b),data=df,
-#'             start=list(Linf=450,k=0.25,ti=3,b=-0.1))
-#' summary(fit3,correlation=TRUE)
-#' curve(rich3(x,Linf=coef(fit3)),from=0,to=15,col="green",lwd=4,add=TRUE)
-#' 
-#' # Fit fourth Richards parameterization
-#' fit4 <- nls(len~rich4(age,Linf,k,ti,b),data=df,
-#'             start=list(Linf=450,k=0.25,ti=3,b=0.7))
-#' summary(fit4,correlation=TRUE)
-#' curve(rich4(x,Linf=coef(fit4)),from=0,to=15,col="black",lwd=1,add=TRUE)
-#' 
-#' ## Logistic
-#' plot(len~age,data=df,pch=19,col=rgb(0,0,0,1/5))
-#' 
-#' # Fit first Campana-Jones parameterization
-#' fit1 <- nls(len~log1(age,Linf,gninf,ti),data=df,
-#'             start=list(Linf=450,gninf=0.45,ti=4))
-#' summary(fit1,correlation=TRUE)
-#' curve(log1(x,Linf=coef(fit1)),from=0,to=15,col="red",lwd=10,add=TRUE)
-#'
-#' # Fit second Campana-Jones parameterization
-#' fit2 <- nls(len~log2(age,Linf,gninf,a),data=df,
-#'             start=list(Linf=450,gninf=0.45,a=7))
-#' summary(fit2,correlation=TRUE)
-#' curve(log2(x,Linf=coef(fit2)),from=0,to=15,col="blue",lwd=5,add=TRUE)
-#'
-#' # Fit Karkach parameterization (using simple=TRUE model)
-#' log3 <- logisticFuns("Karkach",simple=TRUE)
-#' fit3 <- nls(len~log3(age,Linf,L0,gninf),data=df,
-#'             start=list(Linf=450,L0=30,gninf=0.45))
-#' summary(fit3,correlation=TRUE)
-#' curve(log3(x,Linf=coef(fit3)[1],L0=coef(fit3)[2],gninf=coef(fit3)[3]),
-#'       from=0,to=15,col="green",lwd=2,add=TRUE)
-#'
 #'
 #' #############################################################################
-#' ## Create expressions of the models
+#' ## Create expressions of VB models
 #' #############################################################################
 #' # Typical von Bertalanffy ... Show as a stand-alone plot
 #' growthFunShow("vonBertalanffy","Typical",plot=TRUE)
@@ -816,122 +763,103 @@ GompertzFuns <- function(param=c("Ricker1","Ricker2","Ricker3",
 #' @rdname growthModels
 #' @export
 RichardsFuns <- function(param=1,simple=FALSE,msg=FALSE) {
-  Richards1 <- function(t,Linf,k=NULL,a=NULL,b=NULL) {
-    if (length(Linf)==4) { k <- Linf[[2]]
-    a <- Linf[[3]]
-    b <- Linf[[4]]
-    Linf <- Linf[[1]] }
-    Linf*(1-a*exp(-k*t))^b
-  }
-  SRichards1 <- function(t,Linf,k,a,b) {
-    Linf*(1-a*exp(-k*t))^b
-  }
-  Richards2 <- function(t,Linf,k=NULL,ti=NULL,b=NULL) {
+  Richards1 <- function(t,Linf,k=NULL,ti=NULL,b1=NULL) { #eqn 5 from Tjorve & Tjorve (2010)
     if (length(Linf)==4) { k <- Linf[[2]]
     ti <- Linf[[3]]
-    b <- Linf[[4]]
+    b1 <- Linf[[4]]
     Linf <- Linf[[1]] }
-    Linf*(1-(1/b)*exp(-k*(t-ti)))^b
+    Linf/((1+b1*exp(-k*(t-ti)))^(1/b1))
   }
-  SRichards2 <- function(t,Linf,k,ti,b) {
-    Linf*(1-(1/b)*exp(-k*(t-ti)))^b
+  SRichards1 <- function(t,Linf,k,ti,b1) {
+    Linf/((1+b1*exp(-k*(t-ti)))^(1/b1))
   }
-  Richards3 <- function(t,Linf,k=NULL,ti=NULL,b=NULL) {
+  Richards2 <- function(t,Linf,k=NULL,t0=NULL,b2=NULL) { #eqn 3(alt) from Tjorve & Tjorve (2010)
     if (length(Linf)==4) { k <- Linf[[2]]
-    ti <- Linf[[3]]
-    b <- Linf[[4]]
+    t0 <- Linf[[3]]
+    b2 <- Linf[[4]]
     Linf <- Linf[[1]] }
-    Linf/((1+b*exp(-k*(t-ti)))^(1/b))
+    Linf/((1+exp(-k*(t-t0)))^(-b2))
   }
-  SRichards3 <- function(t,Linf,k,ti,b) {
-    Linf/((1+b*exp(-k*(t-ti)))^(1/b))
+  SRichards2 <- function(t,Linf,k,t0,b2) {
+    Linf/((1+exp(-k*(t-t0)))^(-b2))
   }
-  Richards4 <- function(t,Linf,k=NULL,ti=NULL,b=NULL) {
-    if (length(Linf)==4) { k <- Linf[[2]]
-    ti <- Linf[[3]]
-    b <- Linf[[4]]
-    Linf <- Linf[[1]] }
-    Linf*(1+(b-1)*exp(-k*(t-ti)))^(1/(1-b))
-  }
-  SRichards4 <- function(t,Linf,k,ti,b) {
-    Linf*(1+(b-1)*exp(-k*(t-ti)))^(1/(1-b))
-  }
-  Richards5 <- function(t,Linf,k=NULL,L0=NULL,b=NULL) {
+  Richards3 <- function(t,Linf,k=NULL,L0=NULL,b3=NULL) { #eqn 7 from Tjorve & Tjorve (2010)
     if (length(Linf)==4) { k <- Linf[[2]]
     L0 <- Linf[[3]]
-    b <- Linf[[4]]
+    b3 <- Linf[[4]]
     Linf <- Linf[[1]] }
-    Linf*(1+(((L0/Linf)^(1-b))-1)*exp(-k*t))^(1/(1-b))
+    Linf*(1+(((L0/Linf)^(1-b3))-1)*exp(-k*t))^(1/(1-b3))
   }
-  SRichards5 <- function(t,Linf,k,L0,b) {
-    Linf*(1+(((L0/Linf)^(1-b))-1)*exp(-k*t))^(1/(1-b))
+  SRichards3 <- function(t,Linf,k,L0,b3) {
+    Linf*(1+(((L0/Linf)^(1-b3))-1)*exp(-k*t))^(1/(1-b3))
   }
-  Richards6 <- function(t,Linf,k=NULL,ti=NULL,Lninf=NULL,b=NULL) {
-    if (length(Linf)==5) { k <- Linf[[2]]
+  Richards4 <- function(t,Linf,k=NULL,ti=NULL,b2=NULL) { #eqn 4 from Tjorve & Tjorve (2010)
+    if (length(Linf)==4) { k <- Linf[[2]]
     ti <- Linf[[3]]
-    Lninf <- Linf[[3]]
-    b <- Linf[[4]]
+    b2 <- Linf[[4]]
     Linf <- Linf[[1]] }
-    Lninf+(Linf-Lninf)*(1+(b-1)*exp(-k*(t-ti)))^(1/(1-b))
+    Linf*(1-(1/b2)*exp(-k*(t-ti)))^b2
   }
-  SRichards6 <- function(t,Linf,k,ti,Lninf,b) {
-    Lninf+(Linf-Lninf)*(1+(b-1)*exp(-k*(t-ti)))^(1/(1-b))
+  SRichards4 <- function(t,Linf,k,ti,b2) {
+    Linf*(1-(1/b2)*exp(-k*(t-ti)))^b2
   }
+  Richards5 <- function(t,Linf,k=NULL,ti=NULL,b3=NULL) { #eqn 6 from Tjorve & Tjorve (2010)
+    if (length(Linf)==4) { k <- Linf[[2]]
+    ti <- Linf[[3]]
+    b3 <- Linf[[4]]
+    Linf <- Linf[[1]] }
+    Linf*(1+(b3-1)*exp(-k*(t-ti)))^(1/(1-b3))
+  }
+  SRichards5 <- function(t,Linf,k,ti,b3) {
+    Linf*(1+(b3-1)*exp(-k*(t-ti)))^(1/(1-b3))
+  }
+
   ## Main function
-  if (!param %in% 1:6) STOP("'param' must be in 1:6.")
+  if (!param %in% 1:5) STOP("'param' must be in 1:5.")
   param <- paste0("Richards",param)
   if (msg) {
     switch(param,
            Richards1= {
              message("You have chosen the '",param,"' parameterization.",
-                     "  E[L|t] = Linf*(1-a*exp(-k*t))^b\n\n",
+                     "  Linf/((1+b1*exp(-k*(t-ti)))^(1/b1))\n\n",
                      "  where Linf = asymptotic mean length\n",
                      "           k = controls the slope at the inflection point\n",
-                     "           a = a dimensionless parameter that controls the x- value of the inflection point\n",
-                     "           b = a dimensionless parameter that controls the y- value of the inflection point\n\n")
+                     "          ti = time/age at the inflection point\n",
+                     "          b1 = dimensionless, controls the y- value of the inflection point\n\n")
            },
            Richards2= {
              message("You have chosen the '",param,"' parameterization.",
-                     "  Linf*(1-(1/b)*exp(-k*(t-ti)))^b\n\n",
+                     "  Linf/((1-*exp(-k*(t-t0)))^(-b2))\n\n",
                      "  where Linf = asymptotic mean length\n",
                      "           k = controls the slope at the inflection point\n",
-                     "          ti = time/age at the inflection point\n",
-                     "           b = a dimensionless parameter that controls the y- value of the inflection point\n\n")
+                     "          t0 = hypothetical time/age when mean length is 0\n",
+                     "          b2 = dimensionless, controls the y- value of the inflection point\n\n")
            },
            Richards3= {
              message("You have chosen the '",param,"' parameterization.",
-                     "  Linf/((1+b*exp(-k*(t-ti)))^(1/b))\n\n",
-                     "  where Linf = asymptotic mean length\n",
-                     "           k = controls the slope at the inflection point\n",
-                     "          ti = time/age at the inflection point\n",
-                     "           b = a dimensionless parameter that controls the y- value of the inflection point\n\n")
-           },
-           Richards4= {
-             message("You have chosen the '",param,"' parameterization.",
-                     "  Linf*(1+(b-1)*exp(-k*(t-ti)))^(1/(1-b))\n\n",
-                     "  where Linf = asymptotic mean length\n",
-                     "           k = controls the slope at the inflection point\n",
-                     "          ti = time/age at the inflection point\n",
-                     "           b = a dimensionless parameter that controls the y- value of the inflection point\n\n")
-           },
-           Richards5= {
-             message("You have chosen the '",param,"' parameterization.",
-                     "  Linf*(1+(((L0/Linf)^(1-b))-1)*exp(-k*t))^(1/(1-b))\n\n",
+                     "  Linf*(1+(((L0/Linf)^(1-b3))-1)*exp(-k*t))^(1/(1-b3))\n\n",
                      "  where Linf = asymptotic mean length\n",
                      "           k = controls the slope at the inflection point\n",
                      "          L0 = mean length at t=0\n",
-                     "           b = a dimensionless parameter that controls the y- value of the inflection point\n\n")
+                     "          b3 = dimensionless, controls the y- value of the inflection point\n\n")
            },
-           Richards6= {
+           Richards4= {
              message("You have chosen the '",param,"' parameterization.",
-                     "  Lninf+(Linf-Lninf)*(1+(b-1)*exp(-k*(t-ti)))^(1/(1-b))\n\n",
-                     "  where Linf = upper asymptotic mean length\n",
+                     "  Linf*(1-(1/b2)*exp(-k*(t-ti)))^b2\n\n",
+                     "  where Linf = asymptotic mean length\n",
                      "           k = controls the slope at the inflection point\n",
-                     "       Lninf = lower asymptotic mean length\n",
                      "          ti = time/age at the inflection point\n",
-                     "           b = a dimensionless parameter that controls the y- value of the inflection point\n\n")
+                     "          b2 = dimensionless, controls the y- value of the inflection point\n\n")
+           },
+           Richards5= {
+             message("You have chosen the '",param,"' parameterization.",
+                     "  Linf*(1+(b3-1)*exp(-k*(t-ti)))^(1/(1-b3))\n\n",
+                     "  where Linf = asymptotic mean length\n",
+                     "           k = controls the slope at the inflection point\n",
+                     "          ti = time/age at the inflection point\n",
+                     "          b3 = dimensionless, controls the y- value of the inflection point\n\n")
            }
-    )
+           )
   }
   if (simple) param <- paste0("S",param)
   get(param)